+++ api/config/routes.rb
@@ -15,11 +15,11 @@ Rails.application.routes.draw do
   resources :careers, only: :index
   resources :groups, except: [:new, :edit] do
     get :members, on: :member
-    delete 'members/:user_id', to: 'groups#remove_member', on: :member
     resources :requests, only: [:create, :index, :update], module: :groups do
       get 'users/:user_id', to: 'requests#show_for_user', on: :collection
     end
   end
+  resources :members, only: :destroy
   resources :subjects, only: [:index, :show]
   resources :users, only: [:show, :update, :destroy] do
     member do
