+++ api/spec/requests/groups_controller_spec.rb
@@ -418,94 +418,9 @@ RSpec.describe GroupsController, type: :request do
     it 'returns JSON containing the creator as the only member' do
       json_response = response.parsed_body
       expect(json_response.size).to eq(3)
-      expect(json_response[0]['id']).to be_a(Integer)
+      expect(json_response[0]['member_id']).to be_a(Integer)
+      expect(json_response[0]['user_id']).to be_a(Integer)
       expect(json_response[0]['email']).to be_a(String)
-    end
-  end
-
-  # Destroy Member
-  describe 'DELETE /groups/:group_id/members/:user_id' do
-    let(:group) { create :group }
-    let(:admin) { create :user }
-    let(:member) { create :user }
-    let(:another_member) { create :user }
-
-    let(:headersAdmin) do
-      post user_session_path, params: { user: { email: admin.email, password: admin.password } }
-      { 'Authorization' => response.headers['Authorization'] }
-    end
-
-    let(:headersMember) do
-      post user_session_path, params: { user: { email: member.email, password: member.password } }
-      { 'Authorization' => response.headers['Authorization'] }
-    end
-
-    before do
-      group.members.create(user: admin, role: 'admin')
-      group.members.create(user: member, role: 'participant')
-      group.members.create(user: another_member, role: 'participant')
-    end
-
-    context "when user is the group's admin" do
-      before do
-        delete "/groups/#{group.id}/members/#{member.id}", headers: headersAdmin
-      end
-
-      it 'deletes the member successfully' do
-        expect(response).to have_http_status(:no_content)
-        group.reload
-        expect(group.members.find_by(user: member)).to be_nil
-      end
-    end
-
-    context "when user is a group's member" do
-      context 'and tries to remove himself' do
-        before do
-          delete "/groups/#{group.id}/members/#{member.id}", headers: headersMember
-        end
-
-        it 'deletes himself successfully' do
-          expect(response).to have_http_status(:no_content)
-          group.reload
-          expect(group.members.find_by(user: member)).to be_nil
-        end
-      end
-
-      context 'and tries to remove another member' do
-        before do
-          delete "/groups/#{group.id}/members/#{another_member.id}", headers:
-        end
-
-        it 'returns unauthorized status' do
-          expect(response).to have_http_status(:unauthorized)
-        end
-
-        it 'does not delete the other member' do
-          expect(group.members.find_by(user: another_member)).not_to be_nil
-        end
-      end
-
-      context 'when member destruction fails' do
-        let(:headers) do
-          post user_session_path, params: { user: { email: admin.email, password: admin.password } }
-          { 'Authorization' => response.headers['Authorization'] }
-        end
-
-        before do
-          allow_any_instance_of(Member).to receive(:destroy).and_return(false)
-          delete "/groups/#{group.id}/members/#{member.id}", headers:
-        end
-
-        it 'returns an unprocessable entity status' do
-          expect(response).to have_http_status(:unprocessable_entity)
-        end
-
-        it 'returns a JSON containing error messages' do
-          json_response = response.parsed_body
-
-          expect(json_response['message']).to eq('El miembro no pudo ser eliminado correctamente')
-        end
-      end
     end
   end
 end
